+++ src/renderer/src/components/message/UserMessageBlock.tsx
@@ -13,9 +13,10 @@ type Props = {
   baseDir: string;
   message: UserMessage;
   allFiles: string[];
+  renderMarkdown: boolean;
 };
 
-export const UserMessageBlock = ({ baseDir, message, allFiles }: Props) => {
+export const UserMessageBlock = ({ baseDir, message, allFiles, renderMarkdown }: Props) => {
   const baseClasses = 'rounded-md p-3 mb-2 max-w-full break-words whitespace-pre-wrap text-xs bg-neutral-850 border border-neutral-800 text-gray-100';
 
   return (
@@ -27,7 +28,7 @@ export const UserMessageBlock = ({ baseDir, message, allFiles }: Props) => {
         {message.mode === 'context' && <AiOutlineFileSearch className="text-neutral-200" size={16} title="Context Mode" />}
         {message.mode === 'agent' && <RiRobot2Line className="text-neutral-200" size={16} title="Agent Mode" />}
       </div>
-      <div className="flex-grow-1 max-w-full overflow-hidden">{parseMessageContent(baseDir, message.content, allFiles)}</div>
+      <div className="flex-grow-1 max-w-full overflow-hidden">{parseMessageContent(baseDir, message.content, allFiles, renderMarkdown)}</div>
       <div className="absolute top-2 right-2">
         <CopyMessageButton content={message.content} className="text-neutral-600 hover:text-neutral-300" />
       </div>